feat: написан тест для контроллера добавления валют
#INTERNSHIP-42

+++ src/main/java/com/example/bootest/service/client/Valute.java
@@ -14,7 +14,7 @@ public class Valute {
   private String charCode;
 
   @JsonProperty("Nominal")
-  private int nominal;
+  private double nominal;
 
   @JsonProperty("Value")
   private double value;
